# openapi.yaml
openapi: 3.0.3

info:
  title: Restaurant API
  version: 1.0.0
  description: Restaurant REST API for the Product 004 Fresher assignment.

servers:
  - url: http://127.0.0.1:5000

tags:
  - name: Menu
  - name: Customers
  - name: Tables
  - name: Reservations
  - name: Orders

paths:

  /menu:
    get:
      tags:
        - Menu
      operationId: listMenu
      summary: List available menu items
      responses:
        '200':
          description: Available menu items grouped by category.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'

  /menu/{id}:
    get:
      tags:
        - Menu
      operationId: getMenuItem
      summary: Get one menu item
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          description: Menu item not found.

  /customers:
    post:
      tags:
        - Customers
      operationId: createCustomer
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid request.
        '409':
          description: Duplicate email.

  /tables:
    get:
      tags:
        - Tables
      operationId: listDiningTables
      summary: List active dining tables
      responses:
        '200':
          description: Active dining tables.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiningTable'

  /reservations:
    post:
      tags:
        - Reservations
      operationId: createReservation
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationCreate'
      responses:
        '201':
          description: Reservation created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Invalid reservation.
        '409':
          description: Table is already reserved for that time.

  /reservations/{id}:
    get:
      tags:
        - Reservations
      operationId: getReservation
      summary: Get a reservation
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Reservation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservation not found.

  /orders:
    post:
      tags:
        - Orders
      operationId: createOrder
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order.
        '404':
          description: Customer or menu item not found.
        '409':
          description: Menu item unavailable.

  /orders/{id}:
    get:
      tags:
        - Orders
      operationId: getOrder
      summary: Get an order
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Complete order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found.

  /orders/{id}/status:
    patch:
      tags:
        - Orders
      operationId: updateOrderStatus
      summary: Update order status
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
      responses:
        '200':
          description: Order status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status.
        '409':
          description: Invalid status transition.
        '404':
          description: Order not found.

  /customers/{id}/orders:
    get:
      tags:
        - Customers
      operationId: listCustomerOrders
      summary: List orders for a customer
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Customer order history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: Customer not found.

components:

  parameters:

    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

  schemas:

    CustomerCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string

    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerCreate'
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              minimum: 1

    DiningTable:
      type: object
      required:
        - id
        - label
        - seats
        - active
      properties:
        id:
          type: integer
          minimum: 1
        label:
          type: string
        seats:
          type: integer
          minimum: 1
        active:
          type: boolean

    ReservationCreate:
      type: object
      required:
        - customer_id
        - dining_table_id
        - reservation_time
        - party_size
      properties:
        customer_id:
          type: integer
          minimum: 1
        dining_table_id:
          type: integer
          minimum: 1
        reservation_time:
          type: string
        party_size:
          type: integer
          minimum: 1

    Reservation:
      allOf:
        - $ref: '#/components/schemas/ReservationCreate'
        - type: object
          required:
            - id
            - status
          properties:
            id:
              type: integer
              minimum: 1
            status:
              type: string

    MenuItem:
      type: object
      required:
        - id
        - category_id
        - name
        - price_cents
        - available
      properties:
        id:
          type: integer
          minimum: 1
        category_id:
          type: integer
          minimum: 1
        name:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          minimum: 0
        available:
          type: boolean

    OrderItemCreate:
      type: object
      required:
        - menu_item_id
        - quantity
      properties:
        menu_item_id:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1

    OrderCreate:
      type: object
      required:
        - items
      properties:
        customer_id:
          type: integer
          minimum: 1
        dining_table_id:
          type: integer
          minimum: 1
        reservation_id:
          type: integer
          minimum: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemCreate'

    OrderItem:
      type: object
      required:
        - id
        - order_id
        - menu_item_id
        - quantity
        - unit_price_cents
        - line_total_cents
      properties:
        id:
          type: integer
          minimum: 1
        order_id:
          type: integer
          minimum: 1
        menu_item_id:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
        unit_price_cents:
          type: integer
          minimum: 0
        line_total_cents:
          type: integer
          minimum: 0

    Order:
      type: object
      required:
        - id
        - status
        - total_cents
        - created_at
        - items
      properties:
        id:
          type: integer
          minimum: 1
        customer_id:
          type: integer
          nullable: true
        dining_table_id:
          type: integer
          nullable: true
        reservation_id:
          type: integer
          nullable: true
        status:
          type: string
          enum:
            - NEW
            - PREPARING
            - READY
            - COMPLETED
            - CANCELLED
        total_cents:
          type: integer
          minimum: 0
        created_at:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - NEW
            - PREPARING
            - READY
            - COMPLETED
            - CANCELLED